requirements.txt
cryptography>=41.0.0
snowflake-connector-python>=3.0.0
boto3>=1.28.0
